% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_nml}
\alias{read_nml}
\title{read in an *.nml file}
\usage{
read_nml(nml_file)
}
\arguments{
\item{nml_file}{a string with the path to an nml file}
}
\value{
an nml list
}
\description{
read in na nml file and create a list.  \cr
}
\examples{
\dontrun{
sim_nml <- read_nml()
print(sim_nml)
}
}
\author{
Jordan S. Read
}
